+++ presentation/src/main/java/team/aliens/dms_android/feature/notice/NoticeDetailsScreen.kt
@@ -20,7 +20,6 @@ import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
 import java.util.UUID
-import team.aliens.design_system.component.toNoticeDate
 import team.aliens.design_system.theme.DormTheme
 import team.aliens.design_system.typography.Body5
 import team.aliens.design_system.typography.Caption
