+++ presentation/src/main/java/team/aliens/dms_android/feature/notice/NoticesScreen.kt
@@ -2,10 +2,16 @@ package team.aliens.dms_android.feature.notice
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.lazy.LazyColumn
+import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Button
 import androidx.compose.material.ButtonDefaults
 import androidx.compose.runtime.Composable
@@ -17,18 +23,22 @@ import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.clip
+import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.NavController
-import team.aliens.design_system.component.Notice
-import team.aliens.design_system.component.NoticeList
 import team.aliens.design_system.extension.Space
+import team.aliens.design_system.modifier.dormClickable
 import team.aliens.design_system.theme.DormTheme
 import team.aliens.design_system.toast.ToastType
 import team.aliens.design_system.typography.Body1
+import team.aliens.design_system.typography.Body3
+import team.aliens.design_system.typography.Body4
 import team.aliens.design_system.typography.ButtonText
+import team.aliens.design_system.typography.OverLine
 import team.aliens.dms_android.common.LocalAvailableFeatures
 import team.aliens.dms_android.constans.Extra
 import team.aliens.dms_android.feature.application.rememberDmsAppState
@@ -65,7 +75,7 @@ internal fun NoticeScreen(
 
     val onOrderButtonClicked = {
         noticesViewModel.onEvent(
-            event = NoticesUiEvent.SetNoticeOrder,
+            event = NoticesUiEvent.FetchNotices,
         )
         orderText = context.getString(
             when (state.value.order) {
@@ -117,7 +127,7 @@ internal fun NoticeScreen(
             onOrderButtonClicked = onOrderButtonClicked,
         )
         Space(space = 8.dp)
-        NoticeList(
+        Notices(
             notices = notices.map { it.toNotice() },
             onClick = { noticeId ->
                 navController.navigate("noticeDetails/${noticeId}")
@@ -155,4 +165,92 @@ private fun OrderButton(
         }
     }
 }
+
+data class Notice(
+    var noticeId: String,
+    var title: String,
+    var createdAt: String,
+)
+
+@Composable
+fun Notices(
+    modifier: Modifier = Modifier,
+    notices: List<Notice>,
+    errorMessage: String = "",
+    onClick: (String) -> Unit,
+) {
+    LazyColumn(
+        contentPadding = PaddingValues(
+            top = 16.dp,
+            bottom = 64.dp,
+        ),
+        verticalArrangement = Arrangement.spacedBy(8.dp),
+    ) {
+        items(notices) { notice ->
+            Notice(
+                notice = notice,
+                onClick = onClick,
+            )
+        }
+        if (notices.isEmpty()) {
+            item {
+                Body3(
+                    text = errorMessage,
+                )
+            }
+        }
+    }
+}
+
+@Composable
+private fun Notice(
+    notice: Notice,
+    onClick: (String) -> Unit,
+) {
+    Box(
+        modifier = Modifier
+            .fillMaxWidth()
+            .shadow(
+                shape = RoundedCornerShape(
+                    size = 6.dp,
+                ),
+                elevation = 6.dp,
+            )
+            .clip(shape = RoundedCornerShape(6.dp))
+            .background(
+                color = DormTheme.colors.surface,
+            )
+            .dormClickable(
+                rippleEnabled = true,
+            ) {
+                onClick(notice.noticeId)
+            },
+        contentAlignment = Alignment.CenterStart,
+    ) {
+        Column(
+            modifier = Modifier.padding(horizontal = 16.dp),
+        ) {
+            Body4(
+                modifier = Modifier.padding(top = 12.dp),
+                text = notice.title,
+            )
+
+            Space(space = 4.dp)
+
+            OverLine(
+                modifier = Modifier.padding(bottom = 12.dp),
+                text = notice.createdAt.toNoticeDate(),
+                color = DormTheme.colors.primaryVariant,
+            )
+        }
+    }
+}
+
+fun String.toNoticeDate() = StringBuilder().apply {
+    append(this@toNoticeDate.split('T')[0])
+    append(" ")
+    append(this@toNoticeDate.split('T')[1].split(':')[0])
+    append(":")
+    append(this@toNoticeDate.split('T')[1].split(':')[1])
+}.toString()
